Import fixes for thread storage refactor
