+++ src/tools/tool_calls.py
@@ -2,7 +2,7 @@ from typing import Dict, Any
 from fastapi import Request
 
 from src.services.mcp.mcp_manager import McpManager
-from src.services.storage.mongodb_storage import get_mongodb_uri
+from src.services.storage.helpers import get_mongodb_uri
 
 # DEPRECATED
