+++ tests/unit_tests/test_thread_storage.py
@@ -1,7 +1,7 @@
 import pytest
 
 import src.services.storage.mongodb_storage as mongo_storage
-from src.services.storage.mongodb_storage import MongoThreadStorage, MONGODB_COLLECTION_NAME
+from src.services.storage.mongodb_storage import ThreadStorage, MONGODB_COLLECTION_NAME
 from src.services.streaming.stream_variants import SVPrompt, SVUser, SVAssistant, SVStreamEnd
 
 
@@ -11,7 +11,7 @@ async def test_save_and_read_thread(monkeypatch, patch_db, GOOD_HEADERS):
         return "topic"
     monkeypatch.setattr(mongo_storage, "summarize_topic", fake_topic, raising=True)
 
-    storage = await MongoThreadStorage.create(vault_url=GOOD_HEADERS["x-freva-vault-url"])
+    storage = await ThreadStorage.create(vault_url=GOOD_HEADERS["x-freva-vault-url"])
 
     tid = "T123"
     user_id = "alice"
